同步更新版本任务在输入新版本时提示当前版本 #719
