+++ app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sillot",
-  "version": "0.33.1630",
+  "version": "0.33.1700",
   "syv": "3.0.14",
   "sypv": "[]",
   "description": "Build Your Eternal Digital Garden",
