+++ docs/starlight/package.json
@@ -1,7 +1,7 @@
 {
   "name": "starlight",
   "type": "module",
-  "version": "0.33.1630",
+  "version": "0.33.1700",
   "scripts": {
     "test": "echo 'Test your sister day by day'",
     "dev": "astro dev",
